Revert "refactor: move deltalake test to sink test"
This reverts commit e2bb6f0aa4388fe0ecde585d3903c76dc87cf867.
