+++ java/connector-node/risingwave-sink-deltalake/src/test/java/com/risingwave/connector/DeltaLakeSinkFactoryTest.java
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.risingwave.connector.deltalake;
+package com.risingwave.connector;
 
-import com.risingwave.connector.DeltaLakeSinkFactory;
 import com.risingwave.connector.api.TableSchema;
 import io.delta.standalone.DeltaLog;
 import io.delta.standalone.Operation;
