--- java/connector-node/risingwave-sink-test/src/test/resources/log4j2.properties
@@ -1,12 +1,0 @@
-rootLogger.level = ERROR
-# declare the appender to use
-appenders = console
-
-# appender properties
-appender.console.type = Console
-appender.console.name = stdout
-appender.console.layout.type = PatternLayout
-appender.console.layout.pattern = %d{yyyy-MM-dd HH:mm:ss,SSS} %-5p [%t] %c{2}:%L - %m%n
-
-rootLogger.appenderRefs = console
-rootLogger.appenderRef.console.ref = stdout
